requirements.txt
# Core ML
torch>=2.2
torchvision>=0.17
diffusers>=0.27
transformers>=4.40
safetensors>=0.4
accelerate>=0.27
xformers>=0.0.25  # optional — enables memory-efficient attention (gracefully skipped if absent)

# LoRA adapter loading (required by diffusers for load_lora_adapter)
peft>=0.18

# Tagger (JTP PILOT2 SigLIP)
timm>=1.0

# API server
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0

# GPU monitoring (provides pynvml)
nvidia-ml-py>=12.535

# Image / numeric
Pillow>=10.0
numpy>=1.24
